python/requirements.txt
pyparrot
pyserial
opencv_python